move views
Move translations/views under integration/transifex
Move app_manager translations views to translations app
