+++ corehq/apps/translations/urls.py
@@ -1,8 +1,9 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
+
 from django.conf.urls import url
 
-from .views import (
+from corehq.apps.translations.integrations.transifex.views import (
     ConvertTranslations,
     PullResource,
     AppTranslations,
